#203 Componenten ..."carts" verplaatst naar card mapje, grading card refactor, dashboardheader refactor, stats section refactor
Overal wat kleine fixes en checks gedaan
