+++ src/lib/components/cards/ComparisonCard.svelte
