+++ src/lib/components/cards/ProgressCard.svelte
@@ -2,24 +2,19 @@
   import CircleGraph from "$lib/components/charts/CircleGraph.svelte";
   import Button from "$lib/components/buttons/Button.svelte";
 
-  // Props van parent component
   export let articles = [];
 </script>
 
-<!-- Progress overview card HTML -->
 <article class="card progress-card">
-  <!-- Card header met title en subtitle -->
   <header class="card-header">
     <h2 id="progress-heading" class="section-title">Progress overview</h2>
     <p class="section-subtitle body-copy">Completion status of your papers</p>
   </header>
 
-  <!-- Graph container met circle graph component -->
   <div class="graph-container">
     <CircleGraph {articles} />
   </div>
 
-  <!-- Card footer met action button -->
   <footer class="card-footer">
     <Button
       variant="primary"
@@ -32,17 +27,13 @@
 </article>
 
 <style>
-  /* Base card styling */
   .card {
     background: var(--background-color-primary);
-    border-radius: 1rem;
-    box-shadow:
-      0 0.25rem 0.75rem hsla(213, 12%, 15%, 0.12),
-      0 0.125rem 0.25rem hsla(213, 12%, 15%, 0.08);
+    border-radius: var(--radius-lg);
+    box-shadow: var(--shadow-md);
     overflow: hidden;
   }
 
-  /* Progress card met flexbox verticale layout */
   .progress-card {
     display: flex;
     flex-direction: column;
@@ -50,16 +41,16 @@
   }
 
   .card-header {
-    padding: 1.5rem 1.5rem 0 1.5rem;
+    padding: 1.5rem 1.5rem 0;
     margin-bottom: 1.5rem;
 
     @media (min-width: 768px) {
-      padding: 2rem 2rem 0 2rem;
+      padding: 2rem 2rem 0;
     }
   }
 
   .section-title {
-    margin: 0 0 0.5rem 0;
+    margin: 0 0 0.5rem;
     line-height: 1.4;
     color: var(--grey-700);
   }
@@ -70,14 +61,13 @@
     color: var(--grey-600);
   }
 
-  /* Graph container flexbox centering */
+  /* https://developer.mozilla.org/en-US/docs/Web/CSS/flex */
   .graph-container {
     padding: 0 1.5rem;
     display: flex;
     justify-content: center;
     align-items: center;
     min-height: 200px;
-    /* https://developer.mozilla.org/en-US/docs/Web/CSS/flex */
     flex: 1;
 
     @media (min-width: 768px) {
@@ -89,7 +79,7 @@
   .card-footer {
     padding: 1.5rem;
     border-top: 1px solid var(--grey-200);
-    margin-top: auto; 
+    margin-top: auto;
 
     @media (min-width: 768px) {
       padding: 2rem;
